+++ os/osi.cpp
@@ -187,7 +187,7 @@ namespace win
          dwa.set_size(dwa.get_count() + 1024);
          if(!EnumProcesses(
             dwa.get_data(), 
-            dwa.get_count() * sizeof(DWORD),
+            (DWORD) (dwa.get_count() * sizeof(DWORD)),
             &cbNeeded))
          {
             return;
